+++ sim/warlock/imp.go
@@ -87,6 +87,12 @@ func (wp *WarlockPet) registerImpFireboltSpell() {
 	manaCost := [8]float64{0, 10, 20, 35, 50, 70, 95, 115}[rank]
 	level := [8]int{0, 1, 8, 18, 28, 38, 48, 58}[rank]
 
+	// Pimgib's Collar: Equip: Increases the damage of your Imp's Firebolt spell by 18.
+	if wp.owner.HasTrinketEquipped(PimgibsCollar) {
+		baseDamage[0] += 18
+		baseDamage[1] += 18
+	}
+
 	improvedImp := []float64{1, 1.15, 1.3}[wp.owner.Talents.ImprovedImp]
 	baseDamage[0] *= improvedImp
 	baseDamage[1] *= improvedImp